perf: speed up the cli by deferring all pydantic imports
New file type_definitions_heavy.py can have expensive type definitions, keep
only the ones that a fast to load in type_definitions.py.
